Fixed 64-bit compilation.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@8956 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/G1Packet.cpp
@@ -348,7 +348,7 @@ void CG1Packet::GGEPWriteRandomCache(CGGEPBlock& pGGEP, LPCTSTR pszID)
 		return;
 
 	while ( (DWORD)pHosts.GetCount() > Settings.Gnutella1.MaxHostsInPongs )
-		pHosts.RemoveAt( GetRandomNum( 0, pHosts.GetCount() - 1 ) );
+		pHosts.RemoveAt( GetRandomNum( 0, (int)pHosts.GetCount() - 1 ) );
 
 	if ( CGGEPItem* pItem = pGGEP.Add( pszID ) )
 	{
